requirements.txt
fastapi
uvicorn[standard]
httpx
python-jose[cryptography]
passlib[bcrypt]
python-multipart
SQLAlchemy
google-generativeai
python-dotenv
xgboost>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
torch>=2.0.0
scikit-learn>=1.0.0
email-validator
pydantic[email]
